requirements.txt
pandas~=2.2.1
torch~=2.3.0
pillow~=10.3.0
torchvision~=0.18.0